Single view released. Sorting bug fixed
